[WIP] Incorrect interpreter semantics. now run optimize-linalg *incorrectly*
We assume we live in pure functional land, and destroy the `Env` object.
Thus, we have no heap/mutation. I need to split the API into a `Env`
that's pure and a `Heap` that's impure; the `Env` holds references to
`Heap` things.

Alternatively, be more cavalier, just `new MemRef()` and move on with
life...