[WIP] fixing up pass to use the idea where we replace case x of ..       with case (MkFoo x) of ...
I strongly believe that `case` should be a terminator --- it seems to
make many transforms much more complex otherwise. On the other hand,
forcing it to be a terminator makes certain types of optimizations much
harder. Maybe I should use the MLTon encoding, where a case is a type of
switch?