add fusing refcount pass. Provides no optimisation
It appears that at least a naively written pass, which attempts to find
a matching pair of decrement(x);...increment(x) instructions within a
basic block does not find anything useful to optimize.

There are a bunch of situations where we have increment(x);...;decrement(x)[notice the flipped order].
But it seems dubious to treat increment followed by
decrement as a no-op. Indeed, unionfind.lean crashes if I attempt to
"optimise" this pattern away into a no-op.